% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{AUCOR_fun}
\alias{AUCOR_fun}
\title{AUCOR}
\usage{
AUCOR_fun(object)
}
\arguments{
\item{object}{object from function "ave_similarites"}
}
\value{
smooth_line fitted smooth line for average of similaries against
   perturbed sizes

AUCOR AUCOR value
}
\description{
Calculate AUCOR
}
\examples{
 bottomly <- RNAseq_data(sel_size = 3)
 ave <- ave_similarities(counts = bottomly$counts_evaluation,
 group = bottomly$group_evaluation, DEmethod = "edgeR")
 res <- AUCOR_fun(ave$ave_sim)
}
