% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{ave_similarities}
\alias{ave_similarities}
\title{Average similaries for each considered perturbed size}
\usage{
ave_similarities(counts, group, DEmethod = c("edgeR", "edgeR_robust",
  "DESeq_glm", "DESeq2", "samr_SAMseq", "EBSeq", "limma_voom"),
  alpha.vec = seq(0.01, 0.1, 0.01), cut.point = 0.05, nrep = 5)
}
\arguments{
\item{counts}{a matrix of raw read counts}

\item{group}{factor that giving the experimental condition for each sample}

\item{DEmethod}{differential expression analysis method: edgeR, edgeR_robust,
DESeq_glm, DESeq2, samr_SAMseq, EBSeq or limma_voom.}

\item{alpha.vec}{numeric vector of perturbed size, default values is
seq(0.01, 0.1, 0.01).}

\item{cut.point}{threshold for adjusted P values, default value is 0.05.}

\item{nrep}{number of repetition for each perturbed size, default value is 5.}
}
\value{
ave_sim matrix of perturbed size and the corresponding average
   similaries.

pval.mat matrix of p-values and adjusted pvalues for the original
   dataset
}
\description{
Calculate average similaries between sets of selected features from the generated
datasets for each perturbed size and the set of DE features from the
original dataset
}
\examples{
 bottomly <- RNAseq_data(sel_size = 3)
 res <- ave_similarities(counts = bottomly$counts_evaluation,
     group = bottomly$group_evaluation, DEmethod = "edgeR")
}
