% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{DESeq2.pfun}
\alias{DESeq2.pfun}
\title{DESeq2}
\usage{
DESeq2.pfun(counts, group)
}
\arguments{
\item{counts}{a matrix of raw read counts}

\item{group}{factor that giving the experimental condition for each sample}
}
\value{
pval vector of p-values

padj vector of adjusted p-values by Benjamini-Hochberg procedure
}
\description{
Implement DESeq2 (Love et al., 2014) with small
modification from
"http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/".
}
\examples{
    bottomly <- RNAseq_data(sel_size = 3)
    res <- DESeq2.pfun(bottomly$counts_evaluation, bottomly$group_evaluation)


}
\keyword{internal}
