% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{disp_norm}
\alias{disp_norm}
\title{Normalization factors, library size and dispersion}
\usage{
disp_norm(counts, group)
}
\arguments{
\item{counts}{a matrix of raw read counts}

\item{group}{factor that giving the experimental condition for each sample}
}
\value{
tag_disp genewise dispersion arameter estimate for each gene for
        the negative binomial model

norm.factors normalization factors by TMM for each replicate

lib.size library size for each replicate
}
\description{
Calculate normalization factors, library size and dispersion by
edgeR_robust
}
\examples{
 bottomly <- RNAseq_data(sel_size = 3)
 counts <- bottomly$counts_evaluation
 group <- bottomly$group_evaluation
 dn <- disp_norm(counts, group)
}
