% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{edgeR_robust.pfun}
\alias{edgeR_robust.pfun}
\title{edgeR_robust}
\usage{
edgeR_robust.pfun(counts, group, prior.df = 10)
}
\arguments{
\item{counts}{a matrix of raw read counts}

\item{group}{factor that giving the experimental condition for each sample}
}
\value{
pval vector of p-values

padj vector of adjusted p-values by Benjamini-Hochberg procedure
}
\description{
Implement edgeR_robust with standard pipeline (Zhou et al., 2014) with small
modification from
"http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/".
}
\examples{
    bottomly <- RNAseq_data(sel_size = 3)
    res <- edgeR_robust.pfun(bottomly$counts_evaluation, bottomly$group_evaluation)

}
\keyword{internal}
