% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{RNAseq_data}
\alias{RNAseq_data}
\title{Bottomly}
\usage{
RNAseq_data(sel_size = 3)
}
\arguments{
\item{sel_size}{number of replicates of each condition for evaluation set,
default value is 3.}
}
\value{
counts_evaluation  matrix of evaluation RNAseq dataset

group_evaluation factors of evaluation RNAseq dataset

design_evaluation design matrix of evaluation RNAseq dataset

counts_verification  matrix of verification RNAseq dataset

group_verification factors of verification RNAseq dataset

design_verification design matrix of verification RNAseq dataset

counts_full  matrix of full RNAseq dataset

group_full factors of full RNAseq dataset

design_full design matrix of full RNAseq dataset
}
\description{
RNA-seq count data from Bottomly et al. (2011) that compares two
genetically homogeneous mice strains, C57BL/6J and DBA/2J. This dataset contains ten
and eleven replicates for each condition.
}
\examples{
bottomly <- RNAseq_data(sel_size = 3)
}
