% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableDE_fun.R
\name{samr_SAMseq.pfun}
\alias{samr_SAMseq.pfun}
\title{SAMseq}
\usage{
samr_SAMseq.pfun(counts, group)
}
\arguments{
\item{counts}{a matrix of raw read counts}

\item{group}{factor that giving the experimental condition for each sample}
}
\value{
pval vector of p-values

padj vector of adjusted p-values by Benjamini-Hochberg procedure
}
\description{
Implement SAMseq with standard pipeline (Li and Tibshirani,2013) with small
modification from
"http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/".
}
\examples{
    bottomly <- RNAseq_data(sel_size = 3)
    res <- samr_SAMseq.pfun(bottomly$counts_evaluation, bottomly$group_evaluation)

}
\keyword{internal}
